requirements.txt
# Framework Web
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Jinja2==3.1.2

# Banco de Dados
SQLAlchemy==2.0.23

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Exportação Excel
openpyxl==3.1.2
pandas==2.1.4

# Utilitários
python-dotenv==1.0.0
click==8.1.7

# Desenvolvimento e Testes
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0

# Logs e Monitoramento
coloredlogs==15.0.1

# Validação
marshmallow==3.20.1

# Async/Threading
celery==5.3.4
redis==5.0.1

# Segurança
Werkzeug==2.3.7